# contracts/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Namma Kasa API",
    "version": "0.1.0",
    "description": "Live tracking of door-to-door waste collection. Errors are RFC 9457 problem+json."
  },
  "servers": [
    {
      "url": "http://localhost:4000/v1",
      "description": "Local"
    }
  ],
  "paths": {
...
    "/admin/operators": {
      "get": {
        "tags": [
          "admin"
        ],
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "List operators with their ward counts",
        "responses": {
          "200": {
            "description": "Operators",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Operator"
                  }
                }
              }
            }
          },
          "403": {
            "$ref": "#/components/responses/Problem"
          }
        }
      },
      "post": {
        "tags": [
          "admin"
        ],
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Create an operator",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": {
                    "type": "string",
                    "minLength": 1,
                    "maxLength": 200
                  },
                  "type": {
                    "type": "string",
                    "enum": [
                      "bbmp",
                      "private"
                    ]
                  },
                  "config": {
                    "type": "object",
                    "properties": {
                      "complaintSlaHours": {
                        "type": "object",
                        "properties": {
                          "missed_pickup": {
                            "type": "integer",
                            "minimum": 1,
                            "maximum": 720
                          },
                          "late": {
                            "type": "integer",
                            "minimum": 1,
                            "maximum": 720
                          },
                          "behavior": {
                            "type": "integer",
                            "minimum": 1,
                            "maximum": 720
                          },
                          "segregation": {
                            "type": "integer",
                            "minimum": 1,
                            "maximum": 720
                          },
                          "other": {
                            "type": "integer",
                            "minimum": 1,
                            "maximum": 720
                          }
                        },
                        "additionalProperties": false,
                        "default": {}
                      },
                      "defaultComplaintSlaHours": {
                        "type": "integer",
                        "minimum": 1,
                        "maximum": 720,
                        "default": 24
                      },
                      "escalateAfterHours": {
                        "type": "integer",
                        "minimum": 1,
                        "maximum": 720,
                        "default": 24
                      },
                      "sahaayaSyncEnabled": {
                        "type": "boolean",
                        "default": false
                      }
                    },
                    "default": {}
                  }
                },
                "required": [
                  "name",
                  "type"
                ]
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Operator"
                }
              }
            }
          },
          "403": {
            "$ref": "#/components/responses/Problem"
          }
        }
      }
    },
    "/admin/operators/{id}": {
      "patch": {
        "tags": [
          "admin"
        ],
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Update or retire an operator",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "required": true
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": {
                    "type": "string",
                    "minLength": 1,
                    "maxLength": 200
                  },
                  "type": {
                    "type": "string",
                    "enum": [
                      "bbmp",
                      "private"
                    ]
                  },
                  "config": {
                    "type": "object",
                    "properties": {
                      "complaintSlaHours": {
                        "type": "object",
                        "properties": {
                          "missed_pickup": {
                            "type": "integer",
                            "minimum": 1,
                            "maximum": 720
                          },
                          "late": {
                            "type": "integer",
                            "minimum": 1,
                            "maximum": 720
                          },
                          "behavior": {
                            "type": "integer",
                            "minimum": 1,
                            "maximum": 720
                          },
                          "segregation": {
                            "type": "integer",
                            "minimum": 1,
                            "maximum": 720
                          },
                          "other": {
                            "type": "integer",
                            "minimum": 1,
                            "maximum": 720
                          }
                        },
                        "additionalProperties": false,
                        "default": {}
                      },
                      "defaultComplaintSlaHours": {
                        "type": "integer",
                        "minimum": 1,
                        "maximum": 720,
                        "default": 24
                      },
                      "escalateAfterHours": {
                        "type": "integer",
                        "minimum": 1,
                        "maximum": 720,
                        "default": 24
                      },
                      "sahaayaSyncEnabled": {
                        "type": "boolean",
                        "default": false
                      }
                    },
                    "default": {}
                  },
                  "status": {
                    "type": "string",
                    "enum": [
                      "active",
                      "retired"
                    ]
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Updated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Operator"
                }
              }
            }
          },
          "409": {
            "$ref": "#/components/responses/Problem"
          }
        }
      }
    },
    "/admin/wards": {
      "get": {
        "tags": [
          "admin"
        ],
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "List wards visible to the caller",
        "responses": {
          "200": {
            "description": "Wards",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Ward"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "admin"
        ],
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Create a ward",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "operatorId": {
                    "type": "string",
                    "format": "uuid"
                  },
                  "name": {
                    "type": "string",
                    "minLength": 1,
                    "maxLength": 200
                  },
                  "wardCode": {
                    "type": "string",
                    "minLength": 1,
                    "maxLength": 40
                  },
                  "cityId": {
                    "type": "string",
                    "default": "blr"
                  },
                  "boundary": {
                    "anyOf": [
                      {
                        "type": "object",
                        "properties": {
                          "type": {
                            "type": "string",
                            "const": "Polygon"
                          },
                          "coordinates": {
                            "type": "array",
                            "items": {
                              "type": "array",
                              "items": {
                                "type": "array",
                                "maxItems": 2,
                                "minItems": 2,
                                "prefixItems": [
                                  {
                                    "type": "number"
                                  },
                                  {
                                    "type": "number"
                                  }
                                ]
                              }
                            },
                            "minItems": 1
                          }
                        },
                        "required": [
                          "type",
                          "coordinates"
                        ]
                      },
                      {
                        "type": "object",
                        "properties": {
                          "type": {
                            "type": "string",
                            "const": "MultiPolygon"
                          },
                          "coordinates": {
                            "type": "array",
                            "items": {
                              "type": "array",
                              "items": {
                                "type": "array",
                                "items": {
                                  "type": "array",
                                  "maxItems": 2,
                                  "minItems": 2,
                                  "prefixItems": [
                                    {
                                      "type": "number"
                                    },
                                    {
                                      "type": "number"
                                    }
                                  ]
                                }
                              }
                            },
                            "minItems": 1
                          }
                        },
                        "required": [
                          "type",
                          "coordinates"
                        ]
                      }
                    ]
                  }
                },
                "required": [
                  "operatorId",
                  "name",
                  "wardCode",
                  "boundary"
                ]
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Ward"
                }
              }
            }
          },
          "409": {
            "$ref": "#/components/responses/Problem"
          }
        }
      }
    },
    "/admin/wards/{wardId}": {
      "get": {
        "tags": [
          "admin"
        ],
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Fetch one ward",
        "parameters": [
          {
            "in": "path",
            "name": "wardId",
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Ward",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Ward"
                }
              }
            }
          },
          "403": {
            "$ref": "#/components/responses/Problem"
          },
          "404": {
            "$ref": "#/components/responses/Problem"
          }
        }
      },
      "patch": {
        "tags": [
          "admin"
        ],
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Update a ward, including reshaping its boundary",
        "parameters": [
          {
            "in": "path",
            "name": "wardId",
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "required": true
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": {
                    "type": "string",
                    "minLength": 1,
                    "maxLength": 200
                  },
                  "wardCode": {
                    "type": "string",
                    "minLength": 1,
                    "maxLength": 40
                  },
                  "cityId": {
                    "type": "string",
                    "default": "blr"
                  },
                  "boundary": {
                    "anyOf": [
                      {
                        "type": "object",
                        "properties": {
                          "type": {
                            "type": "string",
                            "const": "Polygon"
                          },
                          "coordinates": {
                            "type": "array",
                            "items": {
                              "type": "array",
                              "items": {
                                "type": "array",
                                "maxItems": 2,
                                "minItems": 2,
                                "prefixItems": [
                                  {
                                    "type": "number"
                                  },
                                  {
                                    "type": "number"
                                  }
                                ]
                              }
                            },
                            "minItems": 1
                          }
                        },
                        "required": [
                          "type",
                          "coordinates"
                        ]
                      },
                      {
                        "type": "object",
                        "properties": {
                          "type": {
                            "type": "string",
                            "const": "MultiPolygon"
                          },
                          "coordinates": {
                            "type": "array",
                            "items": {
                              "type": "array",
                              "items": {
                                "type": "array",
                                "items": {
                                  "type": "array",
                                  "maxItems": 2,
                                  "minItems": 2,
                                  "prefixItems": [
                                    {
                                      "type": "number"
                                    },
                                    {
                                      "type": "number"
                                    }
                                  ]
                                }
                              }
                            },
                            "minItems": 1
                          }
                        },
                        "required": [
                          "type",
                          "coordinates"
                        ]
                      }
                    ]
                  },
                  "status": {
                    "type": "string",
                    "enum": [
                      "active",
                      "retired"
                    ]
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Updated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Ward"
                }
              }
            }
          },
          "409": {
            "$ref": "#/components/responses/Problem"
          }
        }
      }
    },
    "/admin/wards/{wardId}/edit-impact": {
      "post": {
        "tags": [
          "admin"
        ],
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Preview what reshaping a boundary would strand",
        "parameters": [
          {
            "in": "path",
            "name": "wardId",
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "required": true
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "anyOf": [
                  {
                    "type": "object",
                    "properties": {
                      "type": {
                        "type": "string",
                        "const": "Polygon"
                      },
                      "coordinates": {
                        "type": "array",
                        "items": {
                          "type": "array",
                          "items": {
                            "type": "array",
                            "maxItems": 2,
                            "minItems": 2,
                            "prefixItems": [
                              {
                                "type": "number"
                              },
                              {
                                "type": "number"
                              }
                            ]
                          }
                        },
                        "minItems": 1
                      }
                    },
                    "required": [
                      "type",
                      "coordinates"
                    ]
                  },
                  {
                    "type": "object",
                    "properties": {
                      "type": {
                        "type": "string",
                        "const": "MultiPolygon"
                      },
                      "coordinates": {
                        "type": "array",
                        "items": {
                          "type": "array",
                          "items": {
                            "type": "array",
                            "items": {
                              "type": "array",
                              "maxItems": 2,
                              "minItems": 2,
                              "prefixItems": [
                                {
                                  "type": "number"
                                },
                                {
                                  "type": "number"
                                }
                              ]
                            }
                          }
                        },
                        "minItems": 1
                      }
                    },
                    "required": [
                      "type",
                      "coordinates"
                    ]
                  }
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Impact",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "affectedHouseholds": {
                      "type": "integer"
                    },
                    "routesOutsideNewBoundary": {
                      "type": "integer"
                    }
                  },
                  "required": [
                    "affectedHouseholds",
                    "routesOutsideNewBoundary"
                  ]
                }
              }
            }
          },
          "404": {
            "$ref": "#/components/responses/Problem"
          }
        }
      }
    },
...
    "/admin/ward-admins": {
      "post": {
        "tags": [
          "admin"
        ],
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Provision a ward admin account",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "wardId": {
                    "type": "string",
                    "format": "uuid"
                  },
                  "phone": {
                    "type": "string",
                    "pattern": "^[1-9]\\d{7,14}$"
                  },
                  "password": {
                    "type": "string",
                    "minLength": 8
                  }
                },
                "required": [
                  "wardId",
                  "phone",
                  "password"
                ]
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "userId": {
                      "type": "string",
                      "format": "uuid"
                    }
                  },
                  "required": [
                    "userId"
                  ]
                }
              }
            }
          },
          "409": {
            "$ref": "#/components/responses/Problem"
          }
        }
      }
    },
...
    "/admin/households/review-queue": {
      "get": {
        "tags": [
          "admin"
        ],
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Households awaiting manual route assignment",
        "responses": {
          "200": {
            "description": "Queue",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ReviewQueueItem"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/admin/households/{id}/route": {
      "patch": {
        "tags": [
          "admin"
        ],
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Place a household on a route by hand",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "required": true
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "routeId": {
                    "type": "string",
                    "format": "uuid"
                  }
                },
                "required": [
                  "routeId"
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Assigned",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "string",
                      "format": "uuid"
                    },
                    "routeId": {
                      "type": "string",
                      "format": "uuid"
                    },
                    "mappingStatus": {
                      "type": "string",
                      "const": "admin_corrected"
                    }
                  },
                  "required": [
                    "id",
                    "routeId",
                    "mappingStatus"
                  ]
                }
              }
            }
          },
          "403": {
            "$ref": "#/components/responses/Problem"
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearer": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
...
      "Operator": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "type": "string"
          },
          "type": {
            "type": "string",
            "enum": [
              "bbmp",
              "private"
            ]
          },
          "config": {
            "type": "object",
            "properties": {
              "complaintSlaHours": {
                "type": "object",
                "properties": {
                  "missed_pickup": {
                    "type": "integer",
                    "minimum": 1,
                    "maximum": 720
                  },
                  "late": {
                    "type": "integer",
                    "minimum": 1,
                    "maximum": 720
                  },
                  "behavior": {
                    "type": "integer",
                    "minimum": 1,
                    "maximum": 720
                  },
                  "segregation": {
                    "type": "integer",
                    "minimum": 1,
                    "maximum": 720
                  },
                  "other": {
                    "type": "integer",
                    "minimum": 1,
                    "maximum": 720
                  }
                },
                "additionalProperties": false,
                "default": {}
              },
              "defaultComplaintSlaHours": {
                "type": "integer",
                "minimum": 1,
                "maximum": 720,
                "default": 24
              },
              "escalateAfterHours": {
                "type": "integer",
                "minimum": 1,
                "maximum": 720,
                "default": 24
              },
              "sahaayaSyncEnabled": {
                "type": "boolean",
                "default": false
              }
            },
            "required": [
              "complaintSlaHours",
              "defaultComplaintSlaHours",
              "escalateAfterHours",
              "sahaayaSyncEnabled"
            ]
          },
          "status": {
            "type": "string",
            "enum": [
              "active",
              "retired"
            ]
          },
          "wardCount": {
            "type": "integer"
          }
        },
        "required": [
          "id",
          "name",
          "type",
          "config",
          "status"
        ]
      },
...
      "ReviewQueueItem": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "fullName": {
            "type": "string"
          },
          "addressLine": {
            "type": "string"
          },
          "landmark": {
            "type": [
              "string",
              "null"
            ]
          },
          "pin": {
            "$ref": "#/components/schemas/GeoPoint"
          },
          "wardId": {
            "type": [
              "string",
              "null"
            ],
            "format": "uuid"
          },
          "createdAt": {
            "type": "string"
          },
          "aging": {
            "type": "boolean"
          }
        },
        "required": [
          "id",
          "fullName",
          "addressLine",
          "landmark",
          "pin",
          "wardId",
          "createdAt",
          "aging"
        ]
      },
...
    },
    "responses": {
      "Problem": {
        "description": "RFC 9457 problem document",
        "content": {
          "application/problem+json": {
            "schema": {
              "$ref": "#/components/schemas/Problem"
            }
          }
        }
      }
    }
  }
}